requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-debug-toolbar==5.1.0
django-environ==0.12.0
gunicorn==23.0.0
kombu==5.5.3
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==6.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
djangorestframework==3.16.0
coverage==7.8.2
django-json-widget==2.0.1